ready for handin

+++ app/src/main/java/com/example/sofie/appointmentmanagementapplication/DeleteChooseFragment.java
@@ -3,16 +3,10 @@ package com.example.sofie.appointmentmanagementapplication;
 import android.app.Fragment;
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
-import android.database.Cursor;
-import android.database.DatabaseUtils;
-import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.EditText;
-import android.widget.TextView;
 
 public class DeleteChooseFragment extends Fragment {
 
@@ -37,19 +31,22 @@ public class DeleteChooseFragment extends Fragment {
             public void onClick(View view) {
                 // get the number of the appointment the user has chosen to delete
                 String siChosenIdToDelete = ChooseAppointmentFragment.etNumberToDelete.getText().toString();
-                int iChosenIdToDelete = Integer.parseInt(siChosenIdToDelete);
+                if(siChosenIdToDelete.trim().length() > 0){
+                    int iChosenIdToDelete = Integer.parseInt(siChosenIdToDelete);
 
-                // replace the DeleteChooseFragment with DeleteConfirmFragment
-                Fragment DeleteConfirmFragment = new DeleteConfirmFragment();
-                // set the variable iChosenIdToDelete in DeleteActivity equal to the id of the appointment the user
-                // wants to delete
-                DeleteActivity.iChosenIdToDelete = ChooseAppointmentFragment.aiAppointmentIds[iChosenIdToDelete - 1];
-                FragmentManager fragmentManager = getActivity().getFragmentManager();
-                FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
+                    // replace the DeleteChooseFragment with DeleteConfirmFragment
+                    Fragment DeleteConfirmFragment = new DeleteConfirmFragment();
+                    // set the variable iChosenIdToDelete in DeleteActivity equal to the id of the appointment the user
+                    // wants to delete
+                    DeleteActivity.iChosenIdToDelete = ChooseAppointmentFragment.aiAppointmentIds[iChosenIdToDelete - 1];
+                    FragmentManager fragmentManager = getActivity().getFragmentManager();
+                    FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
 
-                fragmentTransaction.replace(R.id.fl_delete_activity, DeleteConfirmFragment);
+                    fragmentTransaction.replace(R.id.fl_delete_activity, DeleteConfirmFragment);
 
-                fragmentTransaction.commit();
+                    fragmentTransaction.commit();
+                }
+
             }
         });
     }
